requirements.txt
# Core ML and AI libraries
tensorflow==2.13.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2

# Computer Vision
opencv-python==4.8.0.74
Pillow==10.0.0

# Web Framework
Flask==2.3.2
Flask-CORS==4.0.0

# API and Communication
openai==0.27.8
twilio==8.5.0
requests==2.31.0

# Database
sqlite3

# Geospatial
geopy==2.3.0
folium==0.14.0

# Authentication and Security
PyJWT==2.8.0

# Data Processing
scipy==1.11.1
joblib==1.3.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Development
pytest==7.4.0
black==23.7.0
flake8==6.0.0